Upgraded CursorInputLogic to be overridable easily as well as add CursorButtonHeld events and fixed bug with retrieving the camera from IProxy refs.

+++ UnityProject/Assets/SpacepuppyUnityFramework/Framework/SPInput/SPInput/IInputSignature.cs
@@ -41,7 +41,7 @@ namespace com.spacepuppy.SPInput
         /// <summary>
         /// Last time 'Down' was signaled (as realTimeSinceStartup)
         /// </summary>
-        float LastDownTime { get; }
+        double LastDownTime { get; }
 
     }
